+++ subsys/bluetooth/controller/CMakeLists.txt
@@ -4,32 +4,57 @@ zephyr_library_sources(
   util/memq.c
   util/mayfly.c
   util/util.c
-  hal/nrf5/cntr.c
-  hal/nrf5/rand.c
-  hal/nrf5/ecb.c
-  hal/nrf5/radio/radio.c
   ticker/ticker.c
-  ll_sw/ctrl.c
-  ll_sw/crypto.c
-  ll_sw/ll.c
-  ll_sw/ll_filter.c
+  ll_sw/ll_addr.c
+  ll_sw/ll_tx_pwr.c
+  crypto/crypto.c
   hci/hci_driver.c
   hci/hci.c
   )
 
-zephyr_library_sources_ifdef(CONFIG_BT_BROADCASTER ll_sw/ll_adv.c)
-zephyr_library_sources_ifdef(CONFIG_BT_OBSERVER    ll_sw/ll_scan.c)
-zephyr_library_sources_ifdef(CONFIG_BT_CENTRAL     ll_sw/ll_master.c)
-zephyr_library_sources_ifdef(CONFIG_BT_CTLR_DTM    ll_sw/ll_test.c)
+if(CONFIG_BT_LL_SW)
+  zephyr_library_sources(
+    ll_sw/ctrl.c
+    ll_sw/ll.c
+    )
+  zephyr_library_sources_ifdef(
+    CONFIG_BT_BROADCASTER
+    ll_sw/ll_adv.c
+    )
+  zephyr_library_sources_ifdef(
+    CONFIG_BT_OBSERVER
+    ll_sw/ll_scan.c
+    )
+  zephyr_library_sources_ifdef(
+    CONFIG_BT_CENTRAL
+    ll_sw/ll_master.c
+    )
+  zephyr_library_sources_ifdef(
+    CONFIG_BT_CTLR_DTM
+    ll_sw/ll_test.c
+    )
+endif()
+
+zephyr_library_sources_ifdef(
+  CONFIG_BT_CTLR_FILTER
+  ll_sw/ll_filter.c
+  )
+
+zephyr_library_sources_ifdef(
+  CONFIG_SOC_FAMILY_NRF5
+  hal/nrf5/cntr.c
+  hal/nrf5/rand.c
+  hal/nrf5/ecb.c
+  hal/nrf5/radio/radio.c
+  )
 
 zephyr_library_include_directories(
   .
   util
   hal
   ticker
-  ll
   include
-)
+  )
 
 zephyr_library_compile_options_ifdef(
   CONFIG_BT_CTLR_FAST_ENC
