+++ subsys/bluetooth/controller/crypto/crypto.c
