+++ subsys/bluetooth/controller/hci/hci.c
@@ -385,7 +385,7 @@ static void read_tx_power_level(struct net_buf *buf, struct net_buf **evt)
 
 	rp = cmd_complete(evt, sizeof(*rp));
 
-	status = ll_tx_power_level_get(handle, type, &rp->tx_power_level);
+	status = ll_tx_pwr_lvl_get(handle, type, &rp->tx_power_level);
 	rp->status = (!status) ? 0x00 : BT_HCI_ERR_UNKNOWN_CONN_ID;
 	rp->handle = sys_cpu_to_le16(handle);
 }
@@ -703,6 +703,7 @@ static void le_set_random_address(struct net_buf *buf, struct net_buf **evt)
 	ccst->status = 0x00;
 }
 
+#if defined(CONFIG_BT_CTLR_FILTER)
 static void le_read_wl_size(struct net_buf *buf, struct net_buf **evt)
 {
 	struct bt_hci_rp_le_read_wl_size *rp;
@@ -744,6 +745,7 @@ static void le_rem_dev_from_wl(struct net_buf *buf, struct net_buf **evt)
 	ccst = cmd_complete(evt, sizeof(*ccst));
 	ccst->status = status;
 }
+#endif /* CONFIG_BT_CTLR_FILTER */
 
 static void le_encrypt(struct net_buf *buf, struct net_buf **evt)
 {
@@ -1367,7 +1369,7 @@ static void le_read_tx_power(struct net_buf *buf, struct net_buf **evt)
 
 	rp = cmd_complete(evt, sizeof(*rp));
 	rp->status = 0x00;
-	ll_tx_power_get(&rp->min_tx_power, &rp->max_tx_power);
+	ll_tx_pwr_get(&rp->min_tx_power, &rp->max_tx_power);
 }
 
 #if defined(CONFIG_BT_CTLR_DTM_HCI)
@@ -1454,6 +1456,7 @@ static int controller_cmd_handle(u16_t  ocf, struct net_buf *cmd,
 		le_set_random_address(cmd, evt);
 		break;
 
+#if defined(CONFIG_BT_CTLR_FILTER)
 	case BT_OCF(BT_HCI_OP_LE_READ_WL_SIZE):
 		le_read_wl_size(cmd, evt);
 		break;
@@ -1469,6 +1472,7 @@ static int controller_cmd_handle(u16_t  ocf, struct net_buf *cmd,
 	case BT_OCF(BT_HCI_OP_LE_REM_DEV_FROM_WL):
 		le_rem_dev_from_wl(cmd, evt);
 		break;
+#endif /* CONFIG_BT_CTLR_FILTER */
 
 	case BT_OCF(BT_HCI_OP_LE_ENCRYPT):
 		le_encrypt(cmd, evt);
