+++ subsys/bluetooth/controller/include/ll.h
@@ -73,26 +73,20 @@ u32_t ll_version_ind_send(u16_t handle);
 u32_t ll_terminate_ind_send(u16_t handle, u8_t reason);
 void ll_timeslice_ticker_id_get(u8_t * const instance_index, u8_t * const user_id);
 u32_t ll_rssi_get(u16_t handle, u8_t *rssi);
-u32_t ll_tx_power_level_get(u16_t handle, u8_t type, s8_t *tx_power_level);
-void ll_tx_power_get(s8_t *min, s8_t *max);
+u32_t ll_tx_pwr_lvl_get(u16_t handle, u8_t type, s8_t *tx_pwr_lvl);
+void ll_tx_pwr_get(s8_t *min, s8_t *max);
 
-#if defined(CONFIG_BT_CTLR_LE_PING)
 u32_t ll_apto_get(u16_t handle, u16_t *apto);
 u32_t ll_apto_set(u16_t handle, u16_t apto);
-#endif /* CONFIG_BT_CTLR_LE_PING */
 
-#if defined(CONFIG_BT_CTLR_DATA_LENGTH)
 u32_t ll_length_req_send(u16_t handle, u16_t tx_octets, u16_t tx_time);
 void ll_length_default_get(u16_t *max_tx_octets, u16_t *max_tx_time);
 u32_t ll_length_default_set(u16_t max_tx_octets, u16_t max_tx_time);
 void ll_length_max_get(u16_t *max_tx_octets, u16_t *max_tx_time,
 		       u16_t *max_rx_octets, u16_t *max_rx_time);
-#endif /* CONFIG_BT_CTLR_DATA_LENGTH */
 
-#if defined(CONFIG_BT_CTLR_PHY)
 u32_t ll_phy_get(u16_t handle, u8_t *tx, u8_t *rx);
 u32_t ll_phy_default_set(u8_t tx, u8_t rx);
 u32_t ll_phy_req_send(u16_t handle, u8_t tx, u8_t flags, u8_t rx);
-#endif /* CONFIG_BT_CTLR_PHY */
 
 #endif /* _LL_H_ */
